improve and refactor wishlist feat

+++ features/wishlist/usecase/wishlist-logic.go
@@ -15,24 +15,31 @@ func New(wishlistRepo wishlist.DataInterface) wishlist.UseCaseInterface {
 	}
 }
 
-func (uc *wishlistUsecase) GetAll(userId string) ([]wishlist.WishlistCore, error) {
+func (uc *wishlistUsecase) GetAll(userId string) ([]wishlist.Core, error) {
 	resp, err := uc.wishlistRepository.SelectAll(userId)
 
 	return resp, err
 }
 
-func (uc *wishlistUsecase) Insert(data wishlist.WishlistCore) (wishlist.WishlistCore, error) {
+func (uc *wishlistUsecase) Insert(data wishlist.Core) (wishlist.Core, error) {
 	if data.GameId == "" {
-		return wishlist.WishlistCore{}, errors.New("game id is required")
+		return wishlist.Core{}, errors.New("game id is required")
 	}
 	response, err := uc.wishlistRepository.Insert(data)
 
 	return response, err
 }
 
-func (uc *wishlistUsecase) Delete(id string) error {
+func (uc *wishlistUsecase) Delete(id string, userId string) error {
 	if id == "" {
 		return errors.New("id is required")
+	}
+	data, errGet := uc.wishlistRepository.SelectById(id)
+	if errGet != nil {
+		return errGet
+	}
+	if data.UserId != userId {
+		return errors.New("unauthorized")
 	}
 	err := uc.wishlistRepository.Delete(id)
 	return err
